+++ eslint.config.mjs
@@ -1,0 +1,3 @@
+import { configs } from '@gossi/config-eslint';
+
+export default configs.ember(import.meta.dirname);