+++ prettier.config.mjs
@@ -1,0 +1,1 @@
+export { default } from '@gossi/config-prettier';