+++ src/template-registry.ts
@@ -1,0 +1,5 @@
+import type LinkHelper from './helpers/link.ts';
+
+export default interface LinkRegistry {
+  link: typeof LinkHelper;
+}