+++ src/test-support/index.ts
@@ -1,0 +1,11 @@
+/**
+ * Test support for `ember-link` containing:
+ *
+ * @see https://buschtoens.github.io/ember-link/testing
+ * @module ember-link/test-support
+ */
+
+export { default as TestInstrumentedLinkManagerService } from './-services/test-instrumented-link-manager.ts';
+export { default as linkFor } from './link-for.ts';
+export { default as setupLink } from './setup-link.ts';
+export { default as TestLink } from './test-link.ts';