+++ src/test-support/setup-link.ts
@@ -1,0 +1,26 @@
+import { assert } from '@ember/debug';
+
+import TestInstrumentedLinkManagerService from './-services/test-instrumented-link-manager.ts';
+
+import type { TestContext } from '@ember/test-helpers';
+
+interface Hooks {
+  beforeEach(callback: () => void): void;
+}
+
+export default function setupLink(hooks: Hooks /* NestedHooks */) {
+  hooks.beforeEach(function (this: TestContext) {
+    assert(
+      'ember-link.setupLink: You have already called `setupLink` once',
+      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+      // @ts-ignore
+      !this.owner.hasRegistration('service:link-manager') ||
+        !(this.owner.lookup('service:link-manager') instanceof TestInstrumentedLinkManagerService)
+    );
+
+    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+    // @ts-ignore
+    this.owner.unregister('service:link-manager');
+    this.owner.register('service:link-manager', TestInstrumentedLinkManagerService);
+  });
+}